backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pandas==2.1.4
pydantic==2.5.3
websockets==12.0
emergentintegrations
